test: refactor running the zeromq tests
