+++ .eslintrc.json
@@ -13,7 +13,7 @@
       }
     },
     {
-      "files": ["**/*.test.ts"],
+      "files": ["**/*.ts"],
       "rules": {
         "import/no-extraneous-dependencies": "off"
       }
