+++ .gitignore
@@ -8,3 +8,4 @@ coverage/
 .turbo/
 .cache/
 .DS_Store
+html/
