+++ package.json
@@ -101,6 +101,7 @@
     "vitest": "3.1.1",
     "vite-plugin-babel": "1.3.0",
     "@vitest/coverage-v8": "3.1.1",
+    "@vitest/ui": "3.1.1",
     "@biomejs/biome": "1.9.4",
     "prettier": "3.5.3",
     "prettier-config-atomic": "4.0.0",
